--- ds/beis_indicators/utils/set_utils.py
@@ -1,4 +1,0 @@
-def set_containment(a, b):
-    i = len(set(a).intersection(set(b)))
-    c = i / len(a)
-    return c
